% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSummaRy.R
\name{imageSummaRy}
\alias{imageSummaRy}
\alias{imagesummary}
\alias{imageSummery}
\alias{ImageSummary}
\title{imageSummaRy}
\usage{
imageSummaRy(input_file, color = "none", threshold = 0)
}
\arguments{
\item{input_file}{A character (image file)}

\item{color}{A character (color of the pixels under consideration)}

\item{threshold}{A number (minimum intensity for counting pixles)}
}
\description{
Main function for image summary
}
\details{
Input should be tif-format.
}
\examples{
\dontrun{
# Obtain all positions of cilia in every z-layer
df_image_information <- imageSummaRy(input_files = "inst/image.tif")
}

}
\author{
Kai Budde
}
