% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSummaRy.R
\name{imageSummaRy}
\alias{imageSummaRy}
\alias{imagesummary}
\alias{imageSummery}
\alias{ImageSummary}
\title{imageSummaRy}
\usage{
imageSummaRy(input_file = NULL, input_directory = NULL,
  output_directory = NULL, threshold = 0)
}
\arguments{
\item{input_file}{A character (image file)}

\item{input_directory}{A character (directory of all images)}

\item{output_directory}{A character (directory where the resulting images
are to be saved)}

\item{threshold}{A number (minimum intensity for counting pixles)}
}
\description{
Main function for image summary
}
\details{
Input should be tif-format. The output is the summary in the
data frame format.
}
\examples{
\dontrun{
# Obtain all positions of cilia in every z-layer
df_image_information <- imageSummaRy(input_files = "inst/image.tif")
}

}
\author{
Kai Budde
}
